highlight tags when you mouseover them
As part of this, I'm turning autoranging back on for the y axis on the
chart. Without that, the Y calculation is too simplistic, and fails to
account for the stacking. But I should (a) try and make the autoranging
provide less buffer (0-10 isn't needed, 0-5 is enough for this example)
and (b) make the bottom area a bit bigger.